Initial Commit for the python files.
p4_classification file is the actual classifier.
p4_preprocessing file includes my implementation for a word trie as well as pre-processing the input (read the IMDB files, remove the common words, calculate the priors for the rest of the words, etc...). Interestingly my implementation of word trie used more memory and was noticeably slower than using the built-in Python dictionary.